+++ ADR_vscode_test/Code_Scheme/Coe_Interface.f90
@@ -5,12 +5,12 @@
 !! ----------------------------------------------!!
 Module Parameters
 implicit none
-    real,parameter:: machine_zero = 1E-12
-    real,parameter:: PI = 3.141592653589793238462643
-    real,parameter:: gamma = 1.4
-    real,parameter:: oga   = 1.0/gamma   !< 1.0/GAMMA
-    real,parameter:: vga   = 0.4         !< (GAMMA - 1.0)
-    real,parameter:: ovga  = 2.5         !< 1.0/(GAMMA - 1.0)
+    real(kind=8),parameter:: machine_zero = 1E-12
+    real(kind=8),parameter:: PI = 3.141592653589793238462643
+    real(kind=8),parameter:: gamma = 1.4
+    real(kind=8),parameter:: oga   = 1.0/gamma   !< 1.0/GAMMA
+    real(kind=8),parameter:: vga   = 0.4         !< (GAMMA - 1.0)
+    real(kind=8),parameter:: ovga  = 2.5         !< 1.0/(GAMMA - 1.0)
     integer,parameter:: ZERO = 0
     integer,parameter::  One = 1
     
@@ -27,32 +27,32 @@ implicit none
     integer,parameter:: S6 = 6
     integer,parameter,dimension(10) :: Drv = (/ 1,2,3,4,5,6,7,8,9,10 /)
     
-    real,parameter:: r12 =  1./11
-    real,parameter:: r22 = 10./11
+    real(kind=8),parameter:: r12 =  1./11
+    real(kind=8),parameter:: r22 = 10./11
     
-    real,parameter:: r13 =   1./111
-    real,parameter:: r23 =  10./111
-    real,parameter:: r33 = 100./111
+    real(kind=8),parameter:: r13 =   1./111
+    real(kind=8),parameter:: r23 =  10./111
+    real(kind=8),parameter:: r33 = 100./111
     
-    real,parameter:: r14 =    1./1111
-    real,parameter:: r24 =   10./1111
-    real,parameter:: r34 =  100./1111
-    real,parameter:: r44 = 1000./1111
+    real(kind=8),parameter:: r14 =    1./1111
+    real(kind=8),parameter:: r24 =   10./1111
+    real(kind=8),parameter:: r34 =  100./1111
+    real(kind=8),parameter:: r44 = 1000./1111
     
-    real,parameter:: r15 =     1./11111
-    real,parameter:: r25 =    10./11111
-    real,parameter:: r35 =   100./11111
-    real,parameter:: r45 =  1000./11111
-    real,parameter:: r55 = 10000./11111
+    real(kind=8),parameter:: r15 =     1./11111
+    real(kind=8),parameter:: r25 =    10./11111
+    real(kind=8),parameter:: r35 =   100./11111
+    real(kind=8),parameter:: r45 =  1000./11111
+    real(kind=8),parameter:: r55 = 10000./11111
     
-    real,parameter:: r16 =      1./111111
-    real,parameter:: r26 =     10./111111
-    real,parameter:: r36 =    100./111111
-    real,parameter:: r46 =   1000./111111
-    real,parameter:: r56 =  10000./111111
-    real,parameter:: r66 = 100000./111111
+    real(kind=8),parameter:: r16 =      1./111111
+    real(kind=8),parameter:: r26 =     10./111111
+    real(kind=8),parameter:: r36 =    100./111111
+    real(kind=8),parameter:: r46 =   1000./111111
+    real(kind=8),parameter:: r56 =  10000./111111
+    real(kind=8),parameter:: r66 = 100000./111111
     
-    real,parameter:: lambda = 1.0
+    real(kind=8),parameter:: lambda = 1.0
     
 End Module
 !! ----------------------------------------------!!
